feat: remove core server request logic to new lib
