+++ lib/shared/config-manager/src/request.ts
@@ -1,5 +1,4 @@
-import { RequestInitWithRetry } from 'fetch-retry'
-import { get } from '@devcycle/js-cloud-server-sdk'
+import { getWithTimeout } from '@devcycle/server-request'
 
 export async function getEnvironmentConfig(
     url: string,
@@ -18,21 +17,4 @@ export async function getEnvironmentConfig(
         },
         requestTimeout,
     )
-}
-
-async function getWithTimeout(
-    url: string,
-    requestConfig: RequestInit | RequestInitWithRetry,
-    timeout: number,
-): Promise<Response> {
-    const controller = new AbortController()
-    const id = setTimeout(() => {
-        controller.abort()
-    }, timeout)
-    const response = await get(url, {
-        ...requestConfig,
-        signal: controller.signal,
-    })
-    clearTimeout(id)
-    return response
 }
