+++ lib/shared/server-request/README.md
@@ -1,0 +1,11 @@
+# server-request
+
+This library was generated with [Nx](https://nx.dev).
+
+## Building
+
+Run `nx build server-request` to build the library.
+
+## Running unit tests
+
+Run `nx test server-request` to execute the unit tests via [Jest](https://jestjs.io).
