+++ lib/shared/server-request/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@devcycle/server-request",
+  "version": "0.0.1",
+  "type": "commonjs"
+}
