+++ tsconfig.base.json
@@ -36,6 +36,9 @@
             "@devcycle/openfeature-nodejs-provider": [
                 "sdk/openfeature-nodejs-provider/src/index.ts"
             ],
+            "@devcycle/server-request": [
+                "lib/shared/server-request/src/index.ts"
+            ],
             "@devcycle/react-client-sdk": ["sdk/react/src/index.ts"],
             "@devcycle/react-native-client-sdk": [
                 "sdk/react-native/src/index.ts"
